// backend-buscapet/src/modules/quiz/domain/models/IQuiz.ts
export interface IQuiz {
  id: string;
  birth_date: Date;
  marital_status: string;
  professional_activity: string;
  address: string;
  complement: string;
  district: string;
  city: string;
  state: string;
  cep: string;
  profile_instragam: string;
  for_who: string;
  why_adopt: string;
  average_life: boolean;
  financial_conditions: boolean;
  user_id: string;
  created_at: Date;
  updated_at: Date;
}